requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
groq==0.15.0
jinja2==3.1.5
pytest==8.3.4
pytest-asyncio==0.25.0
email-validator>=2.0.0
streamlit==1.41.1
python-jobspy
